+++ panther/main.py
@@ -2,14 +2,13 @@ import asyncio
 import sys
 import types
 from pathlib import Path
-from rich.console import Console
 from threading import Thread
 
 from panther import status
 from panther._load_configs import *
 from panther._utils import clean_traceback_message, http_response
 from panther.configs import config
-from panther.exceptions import APIException
+from panther.exceptions import APIException, PantherException
 from panther.middlewares.monitoring import Middleware as MonitoringMiddleware
 from panther.request import Request
 from panther.response import Response
@@ -31,7 +30,11 @@ class Panther:
 
         try:
             self.load_configs()
-        except TypeError:
+        except Exception as e:
+            if isinstance(e, PantherException):
+                logger.error(e.args[0])
+            else:
+                logger.error(clean_traceback_message(e))
             exit()
 
         Thread(target=self.websocket_connections, daemon=True, args=(self.ws_redis_connection,)).start()
@@ -42,8 +45,6 @@ class Panther:
 
         # Check & Read The Configs File
         self.configs = load_configs_file(self._configs)
-
-        self.console = Console()
 
         # Put Variables In "config" (Careful about the ordering)
         config['secret_key'] = load_secret_key(self.configs)