+++ docs/docs/release_notes.md
@@ -1,7 +1,16 @@
+### 3.0.3
+- Add `find_one_or_raise` query
+- Add `last_login` to `BaseUser`
+- Add `refresh_life_time` to `JWTConfig`
+- Add `encode_refresh_token()` to `JWTAuthentication`
+- Add `encrypt_password()`
+- Handle `PantherException`
+- Handle `RedisConnection` without `connection_pool`
+
 ### 3.0.2
 - Added 'utf-8' encoding while opening the file "README.md" in setup.py
 - Fixed panther shell not working issue in windows.
-- Added a condition to raise error if no arguement is passed to panther command in cli.
+- Added a condition to raise error if no argument is passed to panther command in cli.
 
 ### 3.0.1
 - Assume content-type is 'application/json' if it was empty